src/requirements.txt
amqp==2.1.3
billiard==3.5.0.2
celery==4.0.0
Django==1.10.1
django-bootstrap-form==3.2.1
django-braces==1.9.0
django-celery==3.1.17
django-cors-headers==1.2.2
django-cors-middleware==1.3.1
django-debug-toolbar==1.6
django-filter==0.15.2
django-memcached==0.1.2
django-oauth-toolkit==0.10.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.4.7
drf-nested-routers==0.11.1
enum==0.4.6
gunicorn==19.6.0
kombu==4.0.1
Markdown==2.6.7
oauth2-provider==0.0
oauthlib==1.0.3
Pillow==3.3.1
PyJWT==1.4.2
python-memcached==1.58
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.10
redis==2.10.5
requests==2.11.1
requests-oauthlib==0.7.0
shortuuid==0.4.3
six==1.10.0
sqlparse==0.2.2
vine==1.1.3
